use 'reclaim' policy for host-path PVs
While the snippet for setting up a PV at the end of this file is only for example usage/testing it needs to use the `Reclaim` policy instead of `Delete` as `Delete` isn't supported for `hostPath` PVs.  Having this set to `Delete` would cause the contents of the PV to *NOT* be deleted even when the PV was deleted which would cause bringing up subsequent iterations of Vault to fail.