fix security context to support specifying any options
At my company, we have a usecase to supply `securityContext.fsGroup: 65534` in order
to satisfy permissions set by amazon for the injection of EKS IAM credentials. Unfortunately,
existing implementation of the chart did not support setting this option.

This change allows any YAML to be specified for `securityContext`, while still leaving existing
configurations the same.