Add VAULT_TOKEN support
If VAULT_TOKEN env variable is defined for the webhook pod,
forward it to all additional init and sidecar containers.
Do not add Vault Agent container if the variable is defined
because looks like it's impossible to reuse an existing token
when using Vault agent.

It's possible to inject secrets into pod, secrets, configmaps
and use Consul templates to generate configuration files
when explicitly specifying VAULT_TOKEN, so you bypass kubernetes
authentication method if using Minikube.