webhook: create sa tokens via apiserver
Currently the webhook tries to find sa.secrets[0]
and load its token. But this is no longer available
from K8S 1.24 (unless users force creating legacy
tokens on each sa which is not best practice). This
change allows creating tokens by calling the apiserver
instead of expecting it to be available as a k8s
secret associated to the service account.