fix: fix vault-env generated env which may contain duplicates
Since https://github.com/banzaicloud/bank-vaults/pull/1697 we are experimenting
issues with `vault-env`: the generated environment has all the `vault:` env
vars duplicated, one version in its original version and a second one with the
secret instead of the token.

I have noticed that the new `preprocessTransitSecrets` is calling `inject` even
 if nothing has been replaced with values from the transit cache. In case you
are working with "maps" (ConfigMap, Secret) this does not matter as the "old"
code will still overwrite the key with the secret. But in the case of
`vault-env` you will have the same env var set twice and most the the time only
 the first occurence (without the secret) is taken into account.