templating support for vault-role setting
Fixes #1160 #954

As described in the mentioned issues, there is a need to define a role based on variables such as the Namespace or ServiceAccount involved.

Instead of manually training developers to set the right annotation, we could define a default using templating to ensure the right role for the right namespace or other concepts.